pipeline/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
numpy>=1.24.0
sentence-transformers>=2.2.2
PyPDF2>=3.0.0 